pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "world-of-shadows-hub"
version = "0.0.0"
description = "Optional editable package: World of Shadows monorepo hub CLIs (Despaghettify, Postmanify, Contractify, …)."
requires-python = ">=3.10"
# Autonomous agents and sandboxes typically run only ``pip install -e .`` here. The hub
# therefore carries the same pinned closure as ``backend/requirements.txt`` plus the
# pytest stack from ``backend/requirements-dev.txt`` and async helpers from
# ``backend/requirements-test.txt``, so ``python tests/run_tests.py --suite backend`` works
# with no second install step. Keep pins aligned with those files (see requirements_hygiene).
dependencies = [
    "alembic>=1.18.0,<2",
    "anyio>=4.0,<5",
    "bleach>=6.0.0,<7",
    "cryptography>=45.0.7,<47",
    "email-validator>=2.0.0,<3",
    'exceptiongroup>=1.2.0; python_version < "3.11"',
    "fastapi>=0.115.0,<1",
    "fastembed>=0.3.0,<1",
    "flask-cors>=4.0.0,<5",
    "flask-jwt-extended>=4.6.0,<5",
    "flask-limiter>=3.5.0,<4",
    "flask-mail>=0.10.0,<1",
    "flask-migrate>=4.0.0,<5",
    "flask-sqlalchemy>=3.1.0,<4",
    "flask-wtf>=1.2.1,<2",
    "flask>=3.0.6,<4",
    "gunicorn>=21.2.0,<23",
    "httpx>=0.27.0,<1",
    "langchain-core>=1.2.0,<2",
    "langchain>=1.2.0,<2",
    "langgraph>=1.0.3,<2",
    "markdown>=3.5.0,<4",
    "marshmallow>=3.20.0,<4",
    "numpy>=1.24.0,<3",
    "pydantic>=2.0.0,<3",
    "pyOpenSSL>=25.3.0,<27",
    "pytest-asyncio>=0.21,<1",
    "pytest-cov>=4.0,<6",
    "pytest-timeout>=2.1",
    "pytest>=7.0,<9",
    "python-dotenv>=1.0.0,<2",
    "pyyaml>=6.0.1,<7",
    "requests>=2.31.0,<3",
    "werkzeug>=3.1.5,<4",
]

[project.scripts]
despag-check = "despaghettify.tools.hub_cli:main"
wos-despag = "despaghettify.tools.hub_cli:main"
postmanify = "postmanify.tools.cli:main"
docify = "docify.tools.hub_cli:main"
contractify = "contractify.tools.hub_cli:main"
documentify = "documentify.tools.hub_cli:main"
testify = "testify.tools.hub_cli:main"
dockerify = "dockerify.tools.hub_cli:main"
fy-platform = "fy_platform.tools.cli:main"

[tool.setuptools.packages.find]
where = ["'fy'-suites"]
